separate TDT and HTK tools
